requirements.txt
oracledb
pillow
torch
open_clip_torch
numpy
dotenv
streamlit